chore: Move util::toml_mut to workspace::editor
