+++ src/workspace/editor/manifest.rs
